+++ tests/tests.py
@@ -6,6 +6,8 @@ import tempfile
 from typing import List, Type, Optional
 from io import StringIO
 
+from unittest.mock import MagicMock
+
 from django.dispatch import receiver  # type: ignore
 from django.test import SimpleTestCase, override_settings  # type: ignore
 
@@ -177,6 +179,38 @@ class FileBasedBackendTests(BaseSmsBackendTests, SimpleTestCase):
             message = message_from_binary_file(fp)
         self.assertEqual(message.originator, '+12065550100')
         self.assertEqual(message.recipients, ['+441134960000'])
+
+
+class MessageBirdBackendTests(BaseSmsBackendTests, SimpleTestCase):
+    sms_backend = 'sms.backends.messagebird.SmsBackend'
+
+    def setUp(self) -> None:
+        super().setUp()
+        self._settings_override = override_settings(
+            MESSAGEBIRD_ACCESS_KEY='fake_access_key'
+        )
+        self._settings_override.enable()
+
+    def tearDown(self) -> None:
+        self._settings_override.disable()
+        super().tearDown()
+
+    def test_send_messages(self) -> None:
+        """Test send_messages with the MessageBird backend."""
+        message = Message(
+            'Here is the message',
+            '+12065550100',
+            ['+441134960000']
+        )
+
+        connection = sms.get_connection()
+        connection.client.message_create = MagicMock()  # type: ignore
+        connection.send_messages([message])  # type: ignore
+        connection.client.message_create.assert_called_with(  # type: ignore
+            '+12065550100',
+            ['+441134960000'],
+            'Here is the message'
+        )
 
 
 class SignalTests(SimpleTestCase):